requirements.txt
google-generativeai
speechrecognition
pyaudio
pygetwindow
pyautogui
python-dotenv
mss
Pillow
